handle exact matches during command suggestions
if `find_suggestable_matches` found a single match it doesn’t need to be
a partial command match. because arguments are left out for each
suggestion step until a suggestion is found, it’s also possible that we
find a single exact match for a command that has been passed to many
arguments. in this case we display the exact command the user probably
wanted to execute.