add event hooks to runner
this adds three events hooks to the runner that can be used to run
arbitrary operations during the events.

on_result
  : every time a result is passed to the console
on_dispatch
  : every time a command is dispatched
on_error
  : every time an error is send to the console

these events are not guarded against exceptions, except those that are
handled internally. if the `--failsafe` switch has been provided
exceptions are intercepted in the application loop.