add support for static mappings in to_enum transformation
if static_map is defined and contains a mapping of enum values to a list
of strings these strings are treated as aliases to the value that is
supposed to be mapped. If any of the strings equals the value the
mapping key it will be used to convert the value to the enum value.