max out terminal window size in kernel
We had customers complain about a weird behaviour that caused
their ECHOed commands to be inserted with a carriage return.

This is enforced by the canonical editor on the terminal itself
to prevent the dot-matrix-printer on the other end from printing
when there is no paper left to print on :).

We couldn’t find a good reason to set the internal window size
to anything else than the maximum value and most likely will
hit a character limit of the canonical editor before having
to worry about the window size again.