allow manual removal of commander provided context objects
Until now the only way to remove a value from the commander context,
that has been provided with `commander.provide('foo', bar)` was to use it
in a `with` statement.

The returned object now allows the removal of the provided value
with the `.destroy()` method, though the use of `.provide(...)`
in a `with`-block is still the preferred way to implement
temporary injection parameters.