fix named injections when used with require
When `kien.commands.require()` is used to generate a
dependency reference it creates an _Injection instance. This
instance is fine-as is and should only be modified when used in a
statement like this

from kien.commands import require
@command.inject(foo=require('bla'))

where `foo` could also be the second argument to `require`.
Unfortunately the implementation replaced `bla` with `foo`
as well, invalidating the `require` statement entirely and
on top of it creating a new _Injection instance for no
particular reason.