add handling for blocking interfaces
When kien wrote to a blocking interface the resulting BlockingIOError
bubbled all the way up to the runner either killing it along with it
or, when using the `--failsafe` option, interrupting the current
running command.

A better option would be for the resulting command to adapt its
behaviour and stop pushing out too much data at the same time.

This can now be implemented by using the new `back_pressure` decorator
in `kien.utils` that utilizes the new CommandExecutionContext feature.