fix: decouple terminal output width handling from help text generation
Even though we do want out help command to provide support for capping
the maximum line length for the generated help text we should not bind
that logic directly to the existence of a terminal. kien should support
use-cases where it is simply used as a text protocol handler and no
terminal is involved whatsoever.

kien now supports an optional `output_width` variable that can be used
to dictate the maximum line length for the generated help. When the
default ConsoleRunner implementation is used it defaults to the terminal
width reported by the blessings library.