fix: fix package installation for systems without preinstalled dependencies
Installing kien with pip only works if the dependencies (currently
blessings and blinker) were already available on the python path. This
is caused by our convenience imports in the `kien` module.

The only reason to import kien during the pip installation was the
__version__ module attribute. As it is easy for us to maintain multiple
version occurences in our codebase through the use of bumpversion we
simply duplicate the __version__ attribute in the setup.py.